// src/types/helpline_row.types.ts
export interface HelplineRowInterface {
  country: string;
  countryI18nKey: string;
  countryCode: string;
  name: string;
  description: string;
  url: string;
  displayUrl: string;
  phone: string;
  displayNumber: string;
  formattedCallNumber: string;
  formattedTextingNumber: string;
  availability: string;
  modality: string;
  identity: string;
  topic: string;
  topicMenu: boolean;
  displayURL?: string;
  twentyFourSeven?: string;
}